requirements.txt
# Core ML dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Utilities
numpy>=1.24.0
requests>=2.31.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Monitoring and logging
loguru>=0.7.0
psutil>=5.9.0

# Queue management and async HTTP
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0

# Additional utilities for queue system
python-dateutil>=2.8.0
